Main Application/requirements.txt

flask
flask_sqlalchemy
qrcode
pyotp
bcrypt
itsdangerous
mysql-connector-python
dotenv
python-multipart
psutil
pyserial
requests
mysql-connector-python
pymysql
mariadb
minimalmodbus
kafka-python
smbus2
pillow
pyodbc
paramiko
pandas
pycomm3
pysmb
python-snap7
RPi.GPIO
pymodbus
dbutils